# Async Crypto Endeavour (+benchmarks!)