--- a/src/etm.js
+++ b/src/etm.js
@@ -4,8 +4,6 @@ const through = require('pull-through')
 const pull = require('pull-stream')
 const lp = require('pull-length-prefixed')
 
-const toForgeBuffer = require('./support').toForgeBuffer
-
 const lpOpts = {
   fixed: true,
   bytes: 4
@@ -13,16 +11,13 @@ const lpOpts = {
 
 exports.createBoxStream = (cipher, mac) => {
   const pt = through(function (chunk) {
-    cipher.update(toForgeBuffer(chunk))
+    const data = cipher.encrypt(chunk)
 
-    if (cipher.output.length() > 0) {
-      const data = new Buffer(cipher.output.getBytes(), 'binary')
-      mac.update(data.toString('binary'))
-      const macBuffer = new Buffer(mac.digest().getBytes(), 'binary')
-
-      this.queue(Buffer.concat([data, macBuffer]))
-      // reset hmac
-      mac.start(null, null)
+    if (data.length > 0) {
+      this.queue(Buffer.concat([
+        data,
+        mac.digest(data)
+      ]))
     }
   })
 
@@ -35,7 +30,7 @@ exports.createBoxStream = (cipher, mac) => {
 exports.createUnboxStream = (decipher, mac) => {
   const pt = through(function (chunk) {
     const l = chunk.length
-    const macSize = mac.getMac().length()
+    const macSize = mac.length
 
     if (l < macSize) {
       return this.emit('error', new Error(`buffer (${l}) shorter than MAC size (${macSize})`))
@@ -45,24 +40,17 @@ exports.createUnboxStream = (decipher, mac) => {
     const data = chunk.slice(0, mark)
     const macd = chunk.slice(mark)
 
-    // Clear out any previous data
-    mac.start(null, null)
+    const expected = mac.digest(data)
 
-    mac.update(data.toString('binary'))
-    const expected = new Buffer(mac.getMac().getBytes(), 'binary')
-
-    // reset hmac
-    mac.start(null, null)
     if (!macd.equals(expected)) {
       return this.emit('error', new Error(`MAC Invalid: ${macd.toString('hex')} != ${expected.toString('hex')}`))
     }
 
     // all good, decrypt
-    decipher.update(toForgeBuffer(data))
+    const encrypted = decipher.decipher(data)
 
-    if (decipher.output.length() > 0) {
-      const data = new Buffer(decipher.output.getBytes(), 'binary')
-      this.queue(data)
+    if (encrypted.length > 0) {
+      this.queue(encrypted)
     }
   })