--- a/src/etm.js
+++ b/src/etm.js
@@ -11,12 +11,22 @@ const lpOpts = {
 
 exports.createBoxStream = (cipher, mac) => {
   const pt = through(function (chunk) {
-    const data = cipher.encrypt(chunk)
+    cipher.encrypt(chunk, (err, data) => {
+      if (err) {
+        return this.emit('error', err)
+      }
 
-    this.queue(Buffer.concat([
-      data,
-      mac.digest(data)
-    ]))
+      mac.digest(data, (err, digest) => {
+        if (err) {
+          return this.emit('error', err)
+        }
+
+        this.queue(Buffer.concat([
+          data,
+          digest
+        ]))
+      })
+    })
   })
 
   return pull(
@@ -38,15 +48,23 @@ exports.createUnboxStream = (decipher, mac) => {
     const data = chunk.slice(0, mark)
     const macd = chunk.slice(mark)
 
-    const expected = mac.digest(data)
+    mac.digest(data, (err, expected) => {
+      if (err) {
+        return this.emit('error', err)
+      }
 
-    if (!macd.equals(expected)) {
-      return this.emit('error', new Error(`MAC Invalid: ${macd.toString('hex')} != ${expected.toString('hex')}`))
-    }
+      if (!macd.equals(expected)) {
+        return this.emit('error', new Error(`MAC Invalid: ${macd.toString('hex')} != ${expected.toString('hex')}`))
+      }
 
-    // all good, decrypt
-    const decrypted = decipher.decipher(data)
-    this.queue(decrypted)
+      // all good, decrypt
+      decipher.decrypt(data, (err, decrypted) => {
+        if (err) {
+          return this.emit('error', err)
+        }
+        this.queue(decrypted)
+      })
+    })
   })
 
   return pull(