--- a/src/handshake/crypto.js
+++ b/src/handshake/crypto.js
@@ -1,15 +1,13 @@
 'use strict'
 
 const protobuf = require('protocol-buffers')
-const path = require('path')
-const fs = require('fs')
 const PeerId = require('peer-id')
 const crypto = require('libp2p-crypto')
 const debug = require('debug')
 const log = debug('libp2p:secio')
 log.error = debug('libp2p:secio:error')
 
-const pbm = protobuf(fs.readFileSync(path.join(__dirname, 'secio.proto')))
+const pbm = protobuf(require('./secio.proto'))
 
 const support = require('../support')
 
@@ -29,7 +27,7 @@ exports.createProposal = (state) => {
   return state.proposalEncoded.out
 }
 
-exports.createExchange = (state) => {
+exports.createExchange = (state, cb) => {
   const res = crypto.generateEphemeralKeyPair(state.protocols.local.curveT)
   state.ephemeralKey.local = res.key
   state.shared.generate = res.genSharedKey
@@ -41,12 +39,18 @@ exports.createExchange = (state) => {
     state.ephemeralKey.local
   ])
 
-  state.exchange.out = {
-    epubkey: state.ephemeralKey.local,
-    signature: new Buffer(state.key.local.sign(selectionOut), 'binary')
-  }
+  state.key.local.sign(selectionOut, (err, sig) => {
+    if (err) {
+      return cb(err)
+    }
 
-  return pbm.Exchange.encode(state.exchange.out)
+    state.exchange.out = {
+      epubkey: state.ephemeralKey.local,
+      signature: sig
+    }
+
+    cb(null, pbm.Exchange.encode(state.exchange.out))
+  })
 }
 
 exports.identify = (state, msg) => {
@@ -62,7 +66,7 @@ exports.identify = (state, msg) => {
   log('1.1 identify - %s - identified remote peer as %s', state.id.local.toB58String(), state.id.remote.toB58String())
 }
 
-exports.selectProtocols = (state) => {
+exports.selectProtocols = (state, cb) => {
   log('1.2 selection')
 
   const local = {
@@ -81,25 +85,30 @@ exports.selectProtocols = (state) => {
     nonce: state.proposal.in.rand
   }
 
-  let selected = support.selectBest(local, remote)
-  // we use the same params for both directions (must choose same curve)
-  // WARNING: if they dont SelectBest the same way, this won't work...
-  state.protocols.remote = {
-    order: selected.order,
-    curveT: selected.curveT,
-    cipherT: selected.cipherT,
-    hashT: selected.hashT
-  }
+  support.selectBest(local, remote, (err, selected) => {
+    if (err) {
+      return cb(err)
+    }
+    // we use the same params for both directions (must choose same curve)
+    // WARNING: if they dont SelectBest the same way, this won't work...
+    state.protocols.remote = {
+      order: selected.order,
+      curveT: selected.curveT,
+      cipherT: selected.cipherT,
+      hashT: selected.hashT
+    }
 
-  state.protocols.local = {
-    order: selected.order,
-    curveT: selected.curveT,
-    cipherT: selected.cipherT,
-    hashT: selected.hashT
-  }
+    state.protocols.local = {
+      order: selected.order,
+      curveT: selected.curveT,
+      cipherT: selected.cipherT,
+      hashT: selected.hashT
+    }
+    cb()
+  })
 }
 
-exports.verify = (state, msg) => {
+exports.verify = (state, msg, cb) => {
   log('2.1. verify')
 
   state.exchange.in = pbm.Exchange.decode(msg)
@@ -111,13 +120,18 @@ exports.verify = (state, msg) => {
     state.ephemeralKey.remote
   ])
 
-  const sigOk = state.key.remote.verify(selectionIn, state.exchange.in.signature)
+  state.key.remote.verify(selectionIn, state.exchange.in.signature, (err, sigOk) => {
+    if (err) {
+      return cb(err)
+    }
 
-  if (!sigOk) {
-    throw new Error('Bad signature')
-  }
+    if (!sigOk) {
+      return cb(new Error('Bad signature'))
+    }
 
-  log('2.1. verify - signature verified')
+    log('2.1. verify - signature verified')
+    cb()
+  })
 }
 
 exports.generateKeys = (state) => {