--- a/src/handshake/propose.js
+++ b/src/handshake/propose.js
@@ -1,8 +1,8 @@
 'use strict'
 
 const debug = require('debug')
-const waterfall = require('run-waterfall')
-
+const waterfall = require('async/waterfall')
+debug.enable('libp2p*')
 const support = require('../support')
 const crypto = require('./crypto')