--- a/src/handshake/secio.proto.js
+++ b/src/handshake/secio.proto.js
@@ -2,15 +2,15 @@
 
 module.exports = new Buffer(`
 message Propose {
-	optional bytes rand = 1;
-	optional bytes pubkey = 2;
-	optional string exchanges = 3;
-	optional string ciphers = 4;
-	optional string hashes = 5;
+  optional bytes rand = 1;
+  optional bytes pubkey = 2;
+  optional string exchanges = 3;
+  optional string ciphers = 4;
+  optional string hashes = 5;
 }
 
 message Exchange {
-	optional bytes epubkey = 1;
-	optional bytes signature = 2;
+  optional bytes epubkey = 1;
+  optional bytes signature = 2;
 }
 `)