--- a/src/handshake/secio.proto.js
+++ b/src/handshake/secio.proto.js
@@ -1,7 +1,6 @@
 'use strict'
 
-module.exports = `
-message Propose {
+module.exports = `message Propose {
   optional bytes rand = 1;
   optional bytes pubkey = 2;
   optional string exchanges = 3;
@@ -12,5 +11,4 @@ message Propose {
 message Exchange {
   optional bytes epubkey = 1;
   optional bytes signature = 2;
-}
-`
+}`