Removes all external crypto libraries, except for those used in `libp2p-crypto` and `multihashing`. This means it drops the usage of node-forge and relies only on webcrypto in the browser. 

Needs https://github.com/libp2p/js-libp2p-crypto/pull/10
## Stats
### Size

|  | non-minified | minified |
| --- | --- | --- |
| before | `1.8M` | `999K` |
| after | `606K` | `382K` |
### Performance

|  | Node.js 6.6.0 | Browser (Chrome) |
| --- | --- | --- |
| before | createKey x 0.40 ops/sec | createKey x 0.31 ops/sec |
|  | send x 0.15 ops/sec | send x 0.11 ops/sec |
| after | createKey x 8.66 ops/sec | createKey x 1.64 ops/sec |
|  | send x 3.61 ops/sec | send x 0.55 ops/sec |