@dignifiedquire 2016-10-04
> There seems to be an issue with go interop on the verification step currently. Investigating currently.

All good, go interop restored in the latest version.